+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Add support for Odyssey's binding preset categories
 - Add Odyssey journal fields
 
+### Changed
+
+- Tolerate malformed graphics configuration files, binding presets
+
 ## [1.12.0](https://github.com/poveden/EliteChroma/compare/v1.11.0...v1.12.0) — 2021-05-24
 
 ### Fixed